modify a class name and merge two files to one file. #2
